Enable time travel for write statements with source queries
