+++ tests/statements/dml.rs
@@ -25,16 +25,18 @@ async fn test_insert_two_different_schemas(
         .unwrap();
     let results = context.collect(plan).await.unwrap();
 
-    let expected = ["+---------------------+------------+------------------+-----------------+----------------+",
+    let expected = [
+        "+---------------------+------------+------------------+-----------------+----------------+",
         "| some_time           | some_value | some_other_value | some_bool_value | some_int_value |",
         "+---------------------+------------+------------------+-----------------+----------------+",
-        "| 2022-01-01T20:01:01 | 42.0       |                  |                 | 1111           |",
-        "| 2022-01-01T20:02:02 | 43.0       |                  |                 | 2222           |",
-        "| 2022-01-01T20:03:03 | 44.0       |                  |                 | 3333           |",
+        "| 2022-01-01T20:01:01 | 42.0       | 1.0000000000     |                 | 1111           |",
+        "| 2022-01-01T20:02:02 | 43.0       | 1.0000000000     |                 | 2222           |",
+        "| 2022-01-01T20:03:03 | 44.0       | 1.0000000000     |                 | 3333           |",
         "|                     | 41.0       | 2.1500000000     | false           |                |",
         "|                     | 45.0       | 9.1200000000     | true            |                |",
         "|                     |            | 44.3400000000    | false           |                |",
-        "+---------------------+------------+------------------+-----------------+----------------+"];
+        "+---------------------+------------+------------------+-----------------+----------------+",
+    ];
     assert_batches_eq!(expected, &results);
 }
 
@@ -376,18 +378,18 @@ async fn test_update_statement(
         "+---------------------+------------+------------------+-----------------+----------------+",
         "| some_time           | some_value | some_other_value | some_bool_value | some_int_value |",
         "+---------------------+------------+------------------+-----------------+----------------+",
-        "| 2022-01-01T21:21:21 | 31.0       |                  |                 | 5555           |",
-        "| 2022-01-01T21:21:21 | 32.0       |                  |                 | 5555           |",
-        "| 2022-01-01T21:21:21 | 32.0       |                  |                 | 5555           |",
-        "| 2022-01-01T21:21:21 | 33.0       |                  |                 | 5555           |",
-        "|                     | 40.0       |                  |                 |                |",
-        "| 2022-01-01T20:03:03 | 44.0       |                  |                 | 3333           |",
-        "|                     | 45.0       |                  |                 |                |",
-        "|                     | 46.0       |                  |                 |                |",
-        "|                     | 46.0       |                  |                 |                |",
-        "|                     | 47.0       |                  |                 |                |",
-        "|                     | 47.0       |                  |                 |                |",
-        "|                     | 48.0       |                  |                 |                |",
+        "| 2022-01-01T21:21:21 | 31.0       | 4.0000000000     |                 | 5555           |",
+        "| 2022-01-01T21:21:21 | 32.0       | 1.0000000000     |                 | 5555           |",
+        "| 2022-01-01T21:21:21 | 32.0       | 4.0000000000     |                 | 5555           |",
+        "| 2022-01-01T21:21:21 | 33.0       | 1.0000000000     |                 | 5555           |",
+        "|                     | 40.0       | 4.0000000000     |                 |                |",
+        "| 2022-01-01T20:03:03 | 44.0       | 1.0000000000     |                 | 3333           |",
+        "|                     | 45.0       | 2.0000000000     |                 |                |",
+        "|                     | 46.0       | 2.0000000000     |                 |                |",
+        "|                     | 46.0       | 3.0000000000     |                 |                |",
+        "|                     | 47.0       | 2.0000000000     |                 |                |",
+        "|                     | 47.0       | 3.0000000000     |                 |                |",
+        "|                     | 48.0       | 3.0000000000     |                 |                |",
         "+---------------------+------------+------------------+-----------------+----------------+",
     ];
     assert_batches_eq!(expected, &results);
